refactor(mills): move enrollment logic out of legacy into enrollment
Behavior-preserving GLIMPSE strangler move (PR-6, zagrajmy/ludamus#457).
Relocate AnonymousEnrollmentService and the user-enrollment-config helpers
from mills/legacy.py into mills/enrollment.py, dropping now-dead imports
from legacy and pointing importers at the new module path.
