+++ src/ludamus/adapters/web/django/forms.py
@@ -20,7 +20,7 @@ from ludamus.adapters.db.django.models import (
     get_used_slots,
     get_vc_available_slots,
 )
-from ludamus.mills import get_user_enrollment_config
+from ludamus.mills.enrollment import get_user_enrollment_config
 from ludamus.pacts import (
     EnrollmentConfigRepositoryProtocol,
     EventDTO,
